Make the runtimes agree with the kernel on equality when comparing doubles and floats
